requirements.txt
streamlit
docxtpl
pillow
streamlit-cropper